perf(bundler): unify AST transforms into one parse/walk pipeline
TreeshakeServerComposables, UseSeoMetaTransform and MinifyTransform each
ran their own oxc parseSync + oxc-walker walk + MagicString over the same
modules, so a module matching all three code filters parsed three times
per build.

The new private createTransformPipeline runs all three concerns over a
single parse, ScopeTracker walk, MagicString and sourcemap:

- rolldown-first filtering: the unified plugin's transform.filter uses the
  union code regex and the shared JS/Vue id filter so rolldown/Vite skip
  the JS hook natively; cheap per-concern regexes re-gate phases inside
  the handler
- phase order per call site is preserved: treeshake claims removed
  statement ranges first, and the seoMeta/minify phases skip any node
  inside a claimed range (previously guaranteed by plugin registration
  order)
- per-target dispatch stays per transform call via resolveBuildConsumer;
  unknown targets retain server code, sharedDuringBuild stays on
- overlapping edits from two phases are detected up front and throw one
  contextual error (file + offsets) with no partial output
- parser failures return undefined (no transform), matching the previous
  MinifyTransform behavior
- the treeshake provenance check accepts importPaths, defaulting to the
  seoMeta importPaths inside the unified pipeline

The public factories (TreeshakeServerComposables, UseSeoMetaTransform,
MinifyTransform) remain exported as thin single-concern adapters over the
pipeline; Unhead()/createFrameworkPlugin() now register one plugin.

Combined transform-positive module: 375 -> 745 ops/s (~2.0x); prefiltered
unrelated module: 333k -> 539k ops/s (no regression).
